requirements.txt
PyExecJS
requests
loguru
python-dotenv
retry
openpyxl
fastapi
uvicorn[standard]
pydantic
pycryptodome